Changes the way log records are cached and processed by the tx committer
The riak_kv_log is changed to send new log records to the tx committer
via the gen_server cast interface.

The ets used to cache log records is removed.

This improves the log record processing latencies because the tx
committer processes new log records as soon as they are available and
not every x ms (pooling ets table).