Optimizes get operations
By discarding vnode snapshots and tagging tentative content with a
tentative version.

If the object does not have tentative contents, the client can just read
the latest version or a version consistent with a snapshot.

If there are tentative contents the client has to wait for them to
become committed if they have a tentative version lower than the
snapshot that was used to read or can just be ignored if their tentative
version is bigger than the snapshot.