Implements the sharded committer
This commit implements the first iteration of the sharded committer.

This first iteration includes a two level commit tree, like:

 Root
  /\
C1  C2

This first iteration is very limitted in terms of the topology that
committers can achieve. It allows the addition of more leaf committers
by just changing the riak configuration file, but does not allow the
addition of more levels in the tree.

In terms of code organisation this commits organises the transaction
logic into four main modules:

- The transactions_manager responsible for managing all the transactions
- The transactions_validator, responsible for validating transactions
(checking if there are any conflicts)
- The transactions_log, responsible for logging the transactions into a
log on disk
- The transactions_committer, responsible for committing the
transactions at the vnodes once a transaction has been validated