Fixes tx manager load balancing bug
Load balancing transactions through the available tx managers was being
done using the vnode idx, but because the idx of a vnode is always an
even number, only even tx managers get transactions.

This commits changes the tx managers to use an hashed idx so that
transactions are distributed more evenly.