Eliminates the need for the tx manager to run inside the riak_kv cluster
This commits creates the riak_core_gateway module which acts as a
gateway to the riak_kv cluster metadata.

With this module the transactions_manager can run outside of the cluster
(in a separate machine for example) and still be able to access the
cluster metadata to commit the transactions.