Fixes the way multi-vnode transactions are validated
Transactions were only validated at the leafs and the validation results
were merged in the upper levels to generate a final validation result.

But this is wrong, because transactions that span different vnodes must
be revalidated again in upper levels of the tree where there is enough
information to do so.

This commit forces transactions that span several vnodes to be
validated again once they get in upper levels of the tree.