Fixes bug in the way conflicts are detected
By giving non-leaf nodes the power to change object versions when
committing multi-vnode transactions, validators can no longer keep only
the latest object version to check if there are conflicts, they must
keep a record of versions along with the transaction that introduced
that version.

This commit changes validators to keep track of the last x versions of
an object along with the trasaction IDs, so that conflitct resolution
can be done correctly.