Updates the way transactional get requests are done
This commit moves the object content selection to the server reducing
the network packet size of get requests.

It also adds logic to the vnodes to block the client until the content
they want to read is committed. The client no longer receives error
messages when the content they want to read is marked as temporary.