Improves garbage collection of object versions at the tx validators
This commit introduces a new object version garbage collection method.

This new method is based on keeping track of object versions from the
last maximum_last_transactions transactions.

Versions created by a transaction with a lsn older than a certain
threshold are deleted.

If the valitor has not enough information to properly validate a
transaction it will conservatively abort the transaction.