Fixes four small bugs
Fixes a bug in the commit fsm that blocked the fsm when more that one
abort message was received. This commit changes the commit fsm to handle
several abort messages for the same transaction.

Fixes a bug that kept the client in a transactional state when an error
was found at commit time. This commit changes the client to get back to
a non transactional state regardless if an error occurs or not.

Fixes a bug that prevented validators from correctly keeping track of
object versions.

Fixes a bug that prevented vnodes from correctly tracking pending
transactional get requests.