Fixes three small bugs
Fixes a bug that kept the client in a transactional state when an error
was found at commit time. This commit changes the client to get back to
a non transactional state regardless if an error occurs or not.

Fixes a bug that prevented vnodes from correctly tracking pending
transactional get requests.

Fixes a bug that prevented tentative versions from getting deleted when
the prepare of a transaction failed.