Adds timeout mechanism to the tx managers batching
This commit changes the way batching is done at the tx managers.

Batching is now done again based on the batch_size parameter. But if the
rate of transactions that the system is processing is slow a new
batch_timeout parameter was added to send the batch of transactions when
the timeout occurs, even if the batch is not full (has reached the
batch_size).